broke reviews into their own component
