+++ src/components/Book.js
@@ -1,196 +1,20 @@
 //=================================================================================================================//
-//                                This version of code goes through book.reviews
-//                              (can add reviews to page but only after page load)
-//                                    (can delete but only after page load)
-//=================================================================================================================//
-// import {useState, useEffect, useCallback} from 'react'
-// import Edit from './Edit.js'
-// import ShowModal from './ShowModal'
-// import ShoppingCart from './ShoppingCart'
-// import axios from 'axios'
-// import AddReview from './AddReview'
-//
-//
-// const Book = (props, book) => {
-//   //--- State:
-//   const [bookData, setBookData] = useState({...props.book})
-//   const[show, setShow] = useState(false)
-//   const [showReviews, setShowReviews] = useState(false)
-//   const [reviews, setReviews] = useState([])
-//   const [bookReviews, setBookReviews] = useState([])
-//
-//
-//   //--- Functions:
-//   //toggles list of book's reviews
-//   const reviewToggle = () => {
-//     if (showReviews === false) {
-//       setShowReviews(true)
-//     } else {
-//       setShowReviews(false)
-//     }
-//   }
-//
-//
-//   //Create Route for reviews
-//   const handleReviewCreate = (addReview) => {
-//    //axios.post('https://ga-bookstore-backend.herokuapp.com/api/books', addBook)
-//    axios.post("http://localhost:8000/api/books/reviews", addReview)
-//    .then((response) => {
-//      setReviews([...reviews, response.data])
-//    })
-//   }
-//
-//   //function the accesses books.reviews
-//   const getBookReviews = () => {
-//     setBookReviews(props.bookReviews)
-//   }
-//
-//   //Delete Route for reviews
-//   const handleReviewDelete = (deletedReview) => {
-//     axios.delete('http://localhost:8000/api/books/reviews/' + deletedReview.id)
-//     .then((response) => {
-//       setReviews(reviews.filter(review => review.id !== deletedReview.id))
-//     })
-//   }
-//
-//   return (
-//       <>
-//         <div className='book' key={bookData.id}>
-//           <img src={bookData.cover_art} alt="book cover"
-//           onClick={() => {
-//             setShow(true)
-//           }}
-//           />
-//           <ShowModal title={bookData.title} onClose={() => setShow(false)} show={show}>
-//           <h2>{bookData.id}</h2>
-//           <img src={bookData.cover_art} alt="book cover"/>
-//           <h5>Author: {bookData.author_name}</h5>
-//           <h5>Publisher: {bookData.publisher}</h5>
-//           <h5>Publication Date: {bookData.publication_date}</h5>
-//           <h5>Pages: {bookData.page_count}</h5>
-//           <h5>Genre: {bookData.genre}</h5>
-//           <h5>Rating: {bookData.rating}</h5>
-//           <br/>
-//           <h5>${bookData.price}</h5>
-//           <input type="number" placeholder="Qty"/>
-//           {props.user === 'admin' ?
-//             <>
-//               <Edit handleUpdate={props.handleUpdate} book={book}/>
-//               <button onClick={() => {props.handleDelete(book)}}>
-//               Delete
-//               </button>
-//             </>
-//           : null}
-//           <div className="all-reviews-div">
-//             <button
-//             onClick={() => {
-//               reviewToggle()
-//               getBookReviews()
-//             }}>See Reviews</button>
-//             {showReviews ?
-//               <>
-//                 <AddReview handleReviewCreate={handleReviewCreate}/>
-//                 <h3>Reviews</h3>
-//                 <div className='all-reviews-flexbox'>
-//                 {bookReviews.map((review) => {
-//                   if (review.book_id === bookData.id) {
-//                     return (
-//                       <div className="review-card" key={review.id}>
-//                         <h5>User: {review.user_id}</h5>
-//                         <h5>Review: {review.review}</h5>
-//                         <h5>review.id: {review.id}</h5>
-//                         <button onClick={() => {handleReviewDelete(review)}}>
-//                         x
-//                         </button>
-//                       </div>
-//                     )
-//                   }
-//                 })}
-//                 </div>
-//               </>
-//             : null}
-//           </div>
-//           </ShowModal>
-//         </div>
-//     </>
-//   )
-// }
-//
-// export default Book
-//=================================================================================================================//
-
-
-
-
-
-//=================================================================================================================//
-//                                This version of code goes through reviews
-//                                             (can delete reviews)
-//                                (cannot add reviews, only through database)
-//                              (added reviews still show up in reviews database)
+//                                This version of code goes through reviews table
 //=================================================================================================================//
 import {useState, useEffect, useCallback} from 'react'
 import Edit from './Edit.js'
 import ShowModal from './ShowModal'
 import ShoppingCart from './ShoppingCart'
 import axios from 'axios'
-import AddReview from './AddReview'
-import EditReview from './EditReview'
+import ReviewsModal from './ReviewsModal'
 
 
 const Book = (props, book) => {
   //--- State:
     const [bookData, setBookData] = useState({...props.book})
     const[show, setShow] = useState(false)
-    const [showReviews, setShowReviews] = useState(false)
-    const [reviews, setReviews] = useState([])
+
     const [bookReviews, setBookReviews] = useState([])
-
-  //--- Functions:
-  const reviewToggle = () => {
-    if (showReviews === false) {
-      setShowReviews(true)
-    } else {
-      setShowReviews(false)
-    }
-  }
-
-  //Create Route for reviews (works only for reviews table)
-  const handleReviewCreate = (addReview) => {
-   //axios.post('https://ga-bookstore-backend.herokuapp.com/api/reviews', addReview)
-   axios.post("http://localhost:8000/api/books/reviews", addReview)
-   .then((response) => {
-     setReviews([...reviews, response.data])
-   })
-  }
-
-  //Read Route for reviews
-  const getBookReviews = () => {
-    //axios.get('https://ga-bookstore-backend.herokuapp.com/api/reviews')
-    axios.get("http://localhost:8000/api/books/reviews")
-    .then(response => setReviews(response.data),
-      err=> console.log(err)
-    ).catch(error=> console.error(error))
-  }
-
-  //Update Route for reviews
-  const handleUpdateReview = (editReview) => {
-    //axios.put('http://localhost:8000/api/books/reviews/' + editReview.id, editReview)
-    axios.put('http://localhost:8000/api/books/reviews/' + editReview.id, editReview)
-    .then((response) => {
-      setReviews(reviews.map((review) => {
-        return review.id !== response.data.id ? review : response.data
-      }))
-    })
-  }
-
-  //Delete Route for reviews
-  const handleReviewDelete = (deletedReview) => {
-    axios.delete('http://localhost:8000/api/books/reviews/' + deletedReview.id)
-    .then((response) => {
-      setReviews(reviews.filter(review => review.id !== deletedReview.id))
-    })
-  }
 
   return (
       <>
@@ -201,7 +25,6 @@ const Book = (props, book) => {
           }}
           />
           <ShowModal title={bookData.title} onClose={() => setShow(false)} show={show}>
-          <h2>{bookData.id}</h2>
           <img src={bookData.cover_art} alt="book cover"/>
           <h5>Author: {bookData.author_name}</h5>
           <h5>Publisher: {bookData.publisher}</h5>
@@ -220,39 +43,7 @@ const Book = (props, book) => {
               </button>
             </>
           : null}
-          <div className="all-reviews-div">
-            <button
-            onClick={() => {
-              reviewToggle()
-              getBookReviews()
-            }}>See Reviews</button>
-            {showReviews ?
-              <>
-                <AddReview handleReviewCreate={handleReviewCreate} bookData={bookData}/>
-                <h3>Reviews</h3>
-                <div className='all-reviews-flexbox'>
-                {reviews.map((review) => {
-                  if (review.book_id === bookData.id) {
-                    return (
-                      <>
-                        <div className="review-card" key={review.id}>
-                           <h5>User: {review.user_id}</h5>
-                           <h5>Review: {review.review}</h5>
-                           <h5>review.id: {review.id}</h5>
-                           <h5>book.id: {review.book_id}</h5>
-                           <EditReview handleUpdateReview={handleUpdateReview} review={review}/>
-                           <button onClick={() => {handleReviewDelete(review)}}>
-                           x
-                           </button>
-                        </div>
-                      </>
-                    )
-                  }
-                })}
-                </div>
-              </>
-            : null}
-          </div>
+          <ReviewsModal bookData={bookData}/>
           </ShowModal>
         </div>
     </>
@@ -587,3 +378,125 @@ export default Book
 // useEffect(() => {
 //   getReviews()
 // }, [])
+
+
+//=================================================================================================================//
+//                                This version of code goes through book.reviews
+//                              (can add reviews to page but only after page load)
+//                                    (can delete but only after page load)
+//=================================================================================================================//
+// import {useState, useEffect, useCallback} from 'react'
+// import Edit from './Edit.js'
+// import ShowModal from './ShowModal'
+// import ShoppingCart from './ShoppingCart'
+// import axios from 'axios'
+// import AddReview from './AddReview'
+//
+//
+// const Book = (props, book) => {
+//   //--- State:
+//   const [bookData, setBookData] = useState({...props.book})
+//   const[show, setShow] = useState(false)
+//   const [showReviews, setShowReviews] = useState(false)
+//   const [reviews, setReviews] = useState([])
+//   const [bookReviews, setBookReviews] = useState([])
+//
+//
+//   //--- Functions:
+//   //toggles list of book's reviews
+//   const reviewToggle = () => {
+//     if (showReviews === false) {
+//       setShowReviews(true)
+//     } else {
+//       setShowReviews(false)
+//     }
+//   }
+//
+//
+//   //Create Route for reviews
+//   const handleReviewCreate = (addReview) => {
+//    //axios.post('https://ga-bookstore-backend.herokuapp.com/api/books', addBook)
+//    axios.post("http://localhost:8000/api/books/reviews", addReview)
+//    .then((response) => {
+//      setReviews([...reviews, response.data])
+//    })
+//   }
+//
+//   //function the accesses books.reviews
+//   const getBookReviews = () => {
+//     setBookReviews(props.bookReviews)
+//   }
+//
+//   //Delete Route for reviews
+//   const handleReviewDelete = (deletedReview) => {
+//     axios.delete('http://localhost:8000/api/books/reviews/' + deletedReview.id)
+//     .then((response) => {
+//       setReviews(reviews.filter(review => review.id !== deletedReview.id))
+//     })
+//   }
+//
+//   return (
+//       <>
+//         <div className='book' key={bookData.id}>
+//           <img src={bookData.cover_art} alt="book cover"
+//           onClick={() => {
+//             setShow(true)
+//           }}
+//           />
+//           <ShowModal title={bookData.title} onClose={() => setShow(false)} show={show}>
+//           <h2>{bookData.id}</h2>
+//           <img src={bookData.cover_art} alt="book cover"/>
+//           <h5>Author: {bookData.author_name}</h5>
+//           <h5>Publisher: {bookData.publisher}</h5>
+//           <h5>Publication Date: {bookData.publication_date}</h5>
+//           <h5>Pages: {bookData.page_count}</h5>
+//           <h5>Genre: {bookData.genre}</h5>
+//           <h5>Rating: {bookData.rating}</h5>
+//           <br/>
+//           <h5>${bookData.price}</h5>
+//           <input type="number" placeholder="Qty"/>
+//           {props.user === 'admin' ?
+//             <>
+//               <Edit handleUpdate={props.handleUpdate} book={book}/>
+//               <button onClick={() => {props.handleDelete(book)}}>
+//               Delete
+//               </button>
+//             </>
+//           : null}
+//           <div className="all-reviews-div">
+//             <button
+//             onClick={() => {
+//               reviewToggle()
+//               getBookReviews()
+//             }}>See Reviews</button>
+//             {showReviews ?
+//               <>
+//                 <AddReview handleReviewCreate={handleReviewCreate}/>
+//                 <h3>Reviews</h3>
+//                 <div className='all-reviews-flexbox'>
+//                 {bookReviews.map((review) => {
+//                   if (review.book_id === bookData.id) {
+//                     return (
+//                       <div className="review-card" key={review.id}>
+//                         <h5>User: {review.user_id}</h5>
+//                         <h5>Review: {review.review}</h5>
+//                         <h5>review.id: {review.id}</h5>
+//                         <button onClick={() => {handleReviewDelete(review)}}>
+//                         x
+//                         </button>
+//                       </div>
+//                     )
+//                   }
+//                 })}
+//                 </div>
+//               </>
+//             : null}
+//           </div>
+//           </ShowModal>
+//         </div>
+//     </>
+//   )
+// }
+//
+// export default Book
+//=================================================================================================================//
